refactor(tests): EIP-7623: minor refactor
