+++ tests/prague/eip7623_increase_calldata_cost/helpers.py
@@ -3,6 +3,7 @@ Helpers for testing EIP-7623.
 """
 
 from enum import Enum, auto
+from typing import Callable
 
 
 class DataTestType(Enum):
@@ -14,11 +15,39 @@ class DataTestType(Enum):
     FLOOR_GAS_COST_GREATER_THAN_INTRINSIC_GAS = auto()
 
 
-class GasTestType(Enum):
+def floor_cost_find(
+    floor_data_gas_cost_calculator: Callable[[int], int],
+    intrinsic_gas_cost_calculator: Callable[[int], int],
+) -> int:
     """
-    Enum for the different types of gas tests.
+    Find the minimum amount of tokens that will trigger the floor gas cost, by using a binary
+    search and the intrinsic gas cost and floor data calculators.
     """
+    # Start with 1000 tokens and if the intrinsic gas cost is greater than the floor gas cost,
+    # multiply the number of tokens by 2 until it's not.
+    tokens = 1000
+    while floor_data_gas_cost_calculator(tokens) < intrinsic_gas_cost_calculator(tokens):
+        tokens *= 2
 
-    CONSUME_ZERO_GAS = auto()
-    CONSUME_ALL_GAS = auto()
-    CONSUME_ALL_GAS_WITH_REFUND = auto()
+    # Binary search to find the minimum number of tokens that will trigger the floor gas cost.
+    left = 0
+    right = tokens
+    while left < right:
+        tokens = (left + right) // 2
+        if floor_data_gas_cost_calculator(tokens) < intrinsic_gas_cost_calculator(tokens):
+            left = tokens + 1
+        else:
+            right = tokens
+    tokens = left
+
+    if floor_data_gas_cost_calculator(tokens) > intrinsic_gas_cost_calculator(tokens):
+        tokens -= 1
+
+    # Verify that increasing the tokens by one would always trigger the floor gas cost.
+    assert (
+        floor_data_gas_cost_calculator(tokens) <= intrinsic_gas_cost_calculator(tokens)
+    ) and floor_data_gas_cost_calculator(tokens + 1) > intrinsic_gas_cost_calculator(
+        tokens + 1
+    ), "invalid case"
+
+    return tokens
